requirements.txt
certifi==2019.11.28
cffi==1.17.1
chardet==3.0.4
coloredlogs==15.0.1
dbus-python==1.2.16
distro-info==0.23+ubuntu1.1
flatbuffers==24.3.25
humanfriendly==10.0
idna==2.8
markdown-it-py==3.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==2.1.2
onnxruntime==1.19.2
packaging==24.1
pika==1.3.2
piper-phonemize==1.1.0
piper-tts==1.2.0
protobuf==5.28.2
pycairo==1.27.0
pycparser==2.22
Pygments==2.18.0
PyGObject==3.36.0
python-apt==2.0.1+ubuntu0.20.4.1
requests==2.22.0
requests-unixsocket==0.2.0
rich==13.9.2
six==1.14.0
sounddevice==0.5.1
sympy==1.13.3
unattended-upgrades==0.1
urllib3==1.25.8
